# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "tiktok clone",
    "description": "\nA simple tiktok clone API\n\n# Sokket.io is used for real-time features like chats, notifications\n\n## chats real-time: ws://localhost:3000/chat?token=<your_token>\n\n### Client -> Server:\n- **Event name**: `SendMessage`\n- **Data**: `{ recipientId: string, message: string }`\n\n### Server -> Client:\n- **Event name**: `NewMessage`\n- **Data**: `{ senderId: string, recipientId: string, message: string }`\n\n\n## Notifications real-time: ws://localhost:3000/notifications?token=<your_token>\n### Server -> Client:\n- **Event name**: `NewNotification`\n- **Data**: `{ userId: string, message: string }`\n\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://course-swap.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/users/{id}/videos": {
      "get": {
        "tags": [
          "UsersController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 10,
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "totalPage": {
                          "type": "number"
                        },
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "title": {
                                "type": "string"
                              },
                              "thumbnail": {
                                "type": "string"
                              },
                              "user": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "username": {
                                    "type": "string"
                                  },
                                  "avatarUrl": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "username",
                                  "avatarUrl"
                                ],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "id",
                              "title",
                              "thumbnail",
                              "user"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPage",
                        "results"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/videos/feed": {
      "get": {
        "tags": [
          "VideosController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "totalPage": {
                          "type": "number"
                        },
                        "results": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "videoUrl": {
                              "type": "string"
                            },
                            "thumbnail": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                            },
                            "likesCount": {
                              "type": "number"
                            },
                            "commentsCount": {
                              "type": "number"
                            },
                            "user": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "username": {
                                  "type": "number"
                                },
                                "avatarUrl": {
                                  "type": "number"
                                }
                              },
                              "required": [
                                "id",
                                "username",
                                "avatarUrl"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "required": [
                            "id",
                            "title",
                            "description",
                            "videoUrl",
                            "thumbnail",
                            "createdAt",
                            "likesCount",
                            "commentsCount",
                            "user"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "limit",
                        "total",
                        "totalPage",
                        "results"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          }
        }
      }
    },
...
    "/api/videos/{id}": {
      "delete": {
        "tags": [
          "VideosController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "UnauthorizedErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/videos/{id}/like": {
      "post": {
        "tags": [
          "LikeController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 200,
                      "maximum": 300
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "data": {
                      "$schema": "https://json-schema.org/draft/2020-12/schema",
                      "type": "object",
                      "properties": {},
                      "additionalProperties": false
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "UnauthorizedErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "http://json-schema.org/draft-07/schema#",
                  "title": "ApiErrorResponse",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "description": "HTTP status code of the error",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable error message"
                    },
                    "name": {
                      "type": "string",
                      "description": "Error class name"
                    }
                  },
                  "required": []
                }
              }
            },
            "description": "Resource not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}